+++ rl_envs_forge/envs/labyrinth/labyrinth.py
@@ -4,6 +4,7 @@ import gymnasium as gym
 import numpy as np
 import random
 import pygame
+import copy
 
 from .maze.maze import MazeFactory
 from .constants import WALL, PATH, TARGET, START, PLAYER, Action
@@ -111,7 +112,24 @@ class Labyrinth(gym.Env):
 
         self.setup_labyrinth()
 
-        self.env_displayer = None  # only initialize if render is needed
+        self._env_displayer = None  # only initialize if render is needed
+        self.displayer_window_size = (800, 800)
+
+    @property
+    def env_displayer(self):
+        if self._env_displayer is None:
+            self._env_displayer = EnvDisplay(
+                self.rows,
+                self.cols,
+                window_width=self.displayer_window_size[0],
+                window_height=self.displayer_window_size[1],
+                labyrinth=self,
+            )
+        return self._env_displayer
+
+    @env_displayer.setter
+    def env_displayer(self, value):
+        self._env_displayer = value
 
     def setup_labyrinth(self) -> None:
         self.make_maze_factory()
@@ -218,6 +236,24 @@ class Labyrinth(gym.Env):
             return False
         return True
 
+    def set_state(self, player_position: Tuple[int, int]) -> None:
+        """Sets the state of the environment. For this environment, we just place the player in a certain position.
+
+        Args:
+            player_position (Tuple[int, int]): The position of the player.
+        """
+        # check that the position is valid
+        if self.maze.grid[player_position[0], player_position[1]] == WALL:
+            raise ValueError("Invalid position, the player can't be on a wall.")
+
+        if np.array_equal(player_position, self.maze.target_position):
+            raise ValueError(
+                "Invalid position, can't place the player on the target position."
+            )
+
+        self.player.position = player_position
+        self.player.rendered_position = player_position
+
     def agent_move(self, action: Action) -> None:
         """
         Move the agent to a new position based on the given action.
@@ -231,7 +267,7 @@ class Labyrinth(gym.Env):
         new_position = self.player.potential_next_position(action)
         self.player.position = new_position
 
-    def reset(self, seed: int = None) -> None:
+    def reset(self, seed: int = None, same_seed: bool = False) -> None:
         """Reset and regenerate another labyrinth. If the same seed as the one at the initialization is provided,
         then the same labyrinth should be regenerated.
 
@@ -242,7 +278,10 @@ class Labyrinth(gym.Env):
         if seed:
             self.seed = seed
         else:
-            self.seed += 1
+            if same_seed:
+                self.seed = self.seed
+            else:
+                self.seed += 1
 
         self.setup_labyrinth()
 
@@ -265,16 +304,7 @@ class Labyrinth(gym.Env):
         Returns:
             Tuple[bool, Action]: A tuple containing a boolean value indicating if the quit event occurred and the action taken by the user.
         """
-        if self.env_displayer is None:
-            # Initialize only if render is needed
-            self.env_displayer = EnvDisplay(
-                self.rows,
-                self.cols,
-                window_width=window_size[0],
-                window_height=window_size[1],
-                labyrinth=self,
-            )
-
+        self.displayer_window_size = window_size
         quit_event = False
         action = None
 
@@ -354,6 +384,7 @@ class Labyrinth(gym.Env):
             )
 
             if quit_event:
+                self.env_displayer = None
                 break
 
             if action is not None:
@@ -380,6 +411,34 @@ class Labyrinth(gym.Env):
             if done:
                 first_info_printed = True
                 self.reset()
+
+    def __deepcopy__(self, memo):
+        # Check if the object is in memo
+        if id(self) in memo:
+            return memo[id(self)]
+
+        # Create a shallow copy of the current environment
+        new_env = copy.copy(self)
+
+        # Add the new environment to memo to avoid infinite loops
+        memo[id(self)] = new_env
+
+        # Manually deep copy attributes that need to be deeply copied
+        new_env.state = np.copy(self.state)
+        new_env.py_random = copy.deepcopy(self.py_random, memo)
+
+        # Manually deep copy the _np_random attribute
+        new_env._np_random = np.random.RandomState()
+        new_env._np_random.set_state(self._np_random.get_state())
+        new_env.reward_schema = copy.deepcopy(self.reward_schema, memo)
+        new_env.player = copy.deepcopy(self.player, memo)
+        new_env.maze = copy.deepcopy(self.maze, memo)
+
+        # Do not deeply copy the env_displayer; set it to None in the copied environment
+        if self._env_displayer is not None:
+            new_env._env_displayer = None
+
+        return new_env
 
 
 if __name__ == "__main__":